feat: IsOwner Resource 리소스에 대해 적용
