+++ src/main/java/com/driveu/server/domain/resource/api/ResourceApi.java
@@ -2,7 +2,6 @@ package com.driveu.server.domain.resource.api;
 
 import com.driveu.server.domain.resource.application.ResourceService;
 import com.driveu.server.domain.resource.dto.request.FileSaveMetaDataRequest;
-import com.driveu.server.domain.resource.dto.request.LinkSaveRequest;
 import com.driveu.server.domain.resource.dto.response.ResourceDeleteResponse;
 import com.driveu.server.domain.resource.dto.response.ResourceFavoriteResponse;
 import com.driveu.server.domain.resource.dto.response.ResourceResponse;
@@ -56,69 +55,6 @@ public class ResourceApi {
         try {
             Long fileId = resourceService.saveFile(user, directoryId, request);
             return ResponseEntity.ok(Map.of("fileId", fileId));
-        } catch (EntityNotFoundException e){
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(Map.of("message", e.getMessage()));
-        } catch (IllegalStateException e){
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("message", e.getMessage()));
-        }
-    }
-
-    @PostMapping("/directories/{directoryId}/links")
-    @Operation(summary = "링크 등록")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "링크 등록 성공",
-                    content = @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(example = "{\"linkId\": \"1\"}")
-                    )),
-            @ApiResponse(responseCode = "404", description = "해당 Directory 없음",
-                    content = @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(example = "{\"message\": \"Directory not found\"}")
-                    )),
-            @ApiResponse(responseCode = "500", description = "서버 내부 오류")
-    })
-    @IsOwner(resourceType = "directory", idParamName = "directoryId")
-    public ResponseEntity<?> uploadLink(
-            @PathVariable Long directoryId,
-            @RequestBody LinkSaveRequest request
-    ) {
-        try {
-            Long linkId = resourceService.saveLink(directoryId, request);
-            return ResponseEntity.ok(Map.of("linkId", linkId));
-        } catch (EntityNotFoundException e){
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(Map.of("message", e.getMessage()));
-        } catch (IllegalStateException e){
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("message", e.getMessage()));
-        }
-    }
-
-    @GetMapping("/links/{linkId}")
-    @Operation(summary = "링크 바로가기")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "링크 바로가기 url 조회 성공",
-                    content = @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(example = "{\"url\": \"https://youtube.com/watch?v=abc123\"}")
-                    )),
-            @ApiResponse(responseCode = "404", description = "해당 Link 없음",
-                    content = @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(example = "{\"message\": \"Link not found\"}")
-                    )),
-            @ApiResponse(responseCode = "500", description = "서버 내부 오류")
-    })
-    public ResponseEntity<?> getLinkUrl(
-            @PathVariable Long linkId,
-            @Parameter(hidden = true) @LoginUser User user
-    ) {
-        try {
-            String url = resourceService.getLinkUrl(linkId);
-            return ResponseEntity.ok(Map.of("url", url));
         } catch (EntityNotFoundException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(Map.of("message", e.getMessage()));
